slap_dj/requirements.txt
Django~=3.1.3
dj_database_url
djangorestframework
django-cors-headers
djangorestframework-csv
drf-nested-routers
django_pandas
coreapi
channels
celery==5.0.2
python-decouple
mysqlclient
pandas
scipy
nltk
plotly

# Scraper and API libraries
lyricsgenius
spotipy
pytrends

